python/python-ai-agents-transition/week-3/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
openai>=1.0.0
tiktoken>=0.5.0